feat: migrate 4 backend files to TypeScript (ADR-010 Phase 1)
Begin backend TS migration per ADR-010 strategy. Migrate the 4 lowest-risk
files (pure functions/constants with no complex require chains) as the
first increment, establishing the migration pattern for subsequent files.

Migrated files:
- src/helpers/providerPresets.ts — pure data (11 AI provider presets)
  Typed: ProviderPresetData interface, ProviderRegistration interface
- src/helpers/ipc-contracts.ts — IPC channel constants (as const)
  Typed: full literal types via 'as const' assertion
- src/helpers/audioPathValidator.ts — path validation function
  Typed: AudioPathValid/AudioPathInvalid discriminated union result type
- src/utils/process.ts — child_process spawn wrapper
  Typed: RunCommandOptions, RunCommandResult interfaces

Migration pattern (reusable for all future files):
1. Create .ts file with proper types (source of truth)
2. Keep .js file as a thin re-export shim: require('./file.ts')
3. Existing require() callers work unchanged via the shim
4. Tests read from .ts when they grep source for patterns
5. Once ALL consumers are .ts, the .js shim can be deleted

Verification: tsc --noEmit clean, 669 tests pass, eslint clean.
